Instalacion de sistema de roles y blindar backend

+++ app/Http/Controllers/Admin/AdminCategoryController.php
@@ -22,6 +22,8 @@ class AdminCategoryController extends Controller
      */
     public function index(Request $request)
     {
+        $this->authorize('haveaccess','category.index');
+
         $nombre = $request->get('nombre');
 
         $categorias = Category::where('nombre', 'like', "%$nombre%")->orderBy('nombre')->paginate(2);
@@ -35,6 +37,8 @@ class AdminCategoryController extends Controller
      */
     public function create()
     {
+        $this->authorize('haveaccess','category.create');
+
         return view('admin.category.create');
     }
 
@@ -56,8 +60,7 @@ class AdminCategoryController extends Controller
         */
 
         //return Category::create($request->all());
-
-
+        $this->authorize('haveaccess','category.create');
         $request->validate([
             'nombre' => 'required|max:50|unique:categories,nombre',
             'slug' => 'required|max:50|unique:categories,slug',
@@ -77,6 +80,7 @@ class AdminCategoryController extends Controller
      */
     public function show($slug)
     {
+        $this->authorize('haveaccess','category.show');
         $cat = Category::where('slug', $slug)->firstOrFail();
         $editar = 'Si';
 
@@ -91,6 +95,7 @@ class AdminCategoryController extends Controller
      */
     public function edit($slug)
     {
+        $this->authorize('haveaccess','category.edit');
         $cat = Category::where('slug', $slug)->firstOrFail();
         $editar = 'Si';
 
@@ -106,8 +111,8 @@ class AdminCategoryController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $this->authorize('haveaccess','category.edit');
         $cat = Category::findOrFail($id);
-
         $request->validate([
             'nombre' => 'required|max:50|unique:categories,nombre,' . $cat->id,
             'slug' => 'required|max:50|unique:categories,slug,' . $cat->id,
@@ -136,6 +141,7 @@ class AdminCategoryController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('haveaccess','category.destroy');
         $cat = Category::findOrFail($id);
         $cat->delete();
         return redirect()->route('admin.category.index')->with('datos', 'Registro eliminado correctamente!');
